Add support for microsecond precision to `time` and `times`
This is an extension of commits c5820aab and b1a41311. This adds
microsecond precision to the times builtin and time keyword. The
precision for both command has been changed to six decimal places
by default.

Additionally, the %C time format from ksh2020 is also added in
this patch. %C is the total number of CPU seconds, i.e. the sum
of %U + %S.