Add support for `stty size` to the libcmd stty builtin
This commit adds support for 'stty size' to the stty builtin, as
defined in https://austingroupbugs.net/view.php?id=1053. The size mode
is used to display the terminal's number of rows and columns. Note that
stty isn't included in the default list of builtin commands; testing
this addition requires adding CMDLIST(stty) to the table of builtins in
src/cmd/ksh93/data/builtins.c.

src/lib/libcmd/stty.c:
- Add support for the size mode to the stty builtin. This mode is only
  used to display the terminal's number of rows and columns, so error
  out if any arguments are given that attempt to set the terminal size.