Use dynamic maximum configuration values when necessary
This commit fixes an issue with how ksh was obtaining the value of
NGROUPS_MAX. On some systems this setting can be changed (e.g., on
illumos adding 'set ngroups_max=32' to /etc/system then rebooting
changes NGROUPS_MAX from 16 to 32). Ksh was using NGROUPS_MAX with
the assumption it's a static value, which could cause issues on systems
where it isn't static. This bugfix is inspired by the one from
<https://github.com/lkujaw/ast/commit/b1362c3a5>, although it has been
expanded a bit to account for OPEN_MAX as well.

getconf(): Rename the macro to 'astgetconf_long' and move it to ast.h to
           prevent redundancy. Other sections of the code have been
	   modified to use this macro for astconf() to account for
	   dynamic settings.