Fix build on illumos (re: a874f8c1, 580ff616)
The standards macros consistency fix for iffe exposed breakage on
illumos: the standards flags aren't set properly. Back in 580ff616,
I set _XPG6 from features/common, which is the wrong place; the
correct place is features/standards -- especially now that iffe
uses its results.

In addition, to get header declarations that aren't somehow in
conflict with themselves on illumos, we need to define _XPG6 *and*
__EXTENSIONS__ *and* _XOPEN_SOURCE. Welp. Thankfully, that's just
fine with Solaris, too.

Thanks to @JohnoKing for the heads-up.