iffe: Correct backslashes in POSIX command substitutions (re: 215efa1)
In backtick command substitutions '\\' will only print one backslash,
while placing a backslash in front of a different character will
print a backslash followed by that character:
   $ echo `echo '\\\\ \3'`
   \\ \3
   $ echo $(echo '\\\\ \3')
   \\\\ \3
This difference in behavior was causing iffe to fail on Haiku after
the changes in 215efa1. This commit applies fixes for backslashes
in quoted strings to fix the Haiku build failure and other possible
bugs introduced by the referenced commit.