Avoid forking subshells when it's unnecessary
The code used to fork subshells when creating/changing aliases will always
fork, even when the alias tree isn't changed:
   $ echo $(unalias --man 2> /dev/null; echo $$ ${.sh.pid})
   375097 375107
   $ alias foo=bar; echo $(alias -p foo; echo $$ ${.sh.pid})
   alias foo=bar 375097 375110
This is a bit inefficient, so this commit applies a minor optimization that
avoid forking a subshell when the alias command was passed -p or -tx and
when the unalias command was passed --help and --man. Note that listing
individual aliases with 'alias foo' will still fork a subshell, as I
haven't changed any of the code for that in this commit.