Fix precision bug in times(3) fallback
This was originally going to be part of a patch for a future feature addition,
but since this fixes a bug I've separated the bugfix part of the patch into its
own pull request.

In the times(3) fallback for the time keyword (which can be enabled in xec.c
by undefining _lib_getrusage and timeofday), ksh will print the obtained
time incorrectly if TIMEFORMAT is set to use a precision level of three:
   $ TIMEFORMAT=$'\nreal\t%3lR'
   $ time sleep .080
   real	0m00.008s  # Should be '00.080s'
This commit corrects that issue by using 10^precision to get the correct
fractional scaling. Note that the fallback still doesn't support a true
precision level of three (times(3) alone doesn't support it), so this
in effect pads a zero to the end of the output when the precision
level is three.

Additional change to tests/builtins.sh:
- While fixing the above issue I found out that ksh93v- broke support for
  passing microseconds to the sleep builtin in the form of <num>U. I've
  added a regression test for that bug to ensure it isn't backported
  to ksh93u+m by accident.