Add NEWS entries for fixed bugs
This commit message will provide explanations for how each bugfix
applied works.

The first bug fixed is the infinite loop that occurs when posix_spawn
fails to execute a command. This was fixed by setting the terminal
process group to the main interactive shell.

The second bug fixed is related to the signal handling of the SIGTTIN,
SIGTTOU and SIGTSTP signals. In sh_ntfork() these signals are set to
their default signal handlers (SIG_DFL) before running a command. The
signal handlers were only restored to SIG_IGN (ignore signal) when
sh_ntfork() successfully ran a command. This could cause a SIGTTOU
lockup under strace when a command failed to execute in an interactive
shell, while also being one cause of the infinite loop.