Fix read -a failure to create array (re: 4d182da8)
The commit that backported 'read -a' did not add a case label for it to
read.c. This was under the assumption that optget would always convert
-a to -A, although this doesn't seem to be very consistent behavior.
After running 'read -a' or 'read -A' once, the -a flag will fail to
to create an array while -A is unaffected. Reproducer:
   $ read -A foo <<< 'foo bar baz'
   $ unset foo
   $ read -a foo <<< 'foo bar baz'
   $ echo ${foo[0]}
   foo bar baz
   $ echo ${foo[1]}   # Prints a single newline and nothing else
   $ read -A foo <<< 'foo bar baz'
   $ echo ${foo[0]}
   foo
   $ echo ${foo[1]}
   bar

src/cmd/ksh93/bltins/read.c:
- Add case label for the read builtin's -a flag.

src/cmd/ksh93/tests/builtins.sh:
- Add regression tests for this bug. For consistency these are all
  run in a separate script.