optget.c: Fix ASan crash when using --version with C.UTF-8
This commit fixes a crash that occurred under ASan in the libcmd
regression tests. In some scenarios using --version to display a
builtin's version will cause a crash because ksh will attempt to read
from one byte before the version string. This is avoided by using a
blank space as the first byte of the string. Backported from ksh2020:
https://github.com/att/ast/commit/a9a9fafa
https://github.com/att/ast/commit/3e73d550