emacs.c: use sh_trap for ESC Ctrl-L again (re: 7ff6b73b, 83630f9d)
Using 'command -p' via system(3) is problematic as 'command -p' is
broken on far too many shells. So invoking a command with 'command
-p' needs to be done with ksh itself, via sh_trap(). Previously
this crashed (see 7ff6b73b) but I think we've fixed that in the
meantime, as it's working fine on my end now (plus, the analysis of
the cause of the crash in that commit was wrong); see 83630f9d.

Also block SIGINT and turn off restricted, verbose and xtrace so
that this works reliably and without terminal noise.