Fix problems in relation to Fedora's C99 instrumentation
src/lib/libast/comp/conf.tab:
- Add include directives to fix implicit function declarations that
  cause the PID_MAX test to silently fail. This compile error occurs
  on all platforms when compiling with strict C99, but wasn't
  exposed even after setting IFFEFLAGS to -d1 or -d2 (that's likely
  a separate bug in conf.sh that'll need fixing). As a consequence,
  it could cause following bug in the getconf builtin (now fixed
  as of this commit):
  $ /opt/ast/bin/getconf PID_MAX
  30000  # Wrong, should be the platform value (e.g., 4194304)

src/lib/libast/{features/lib,sfio/sfhdr.h}:
- Delete the lib_poll_fd_2 feature test and associated code, which
  allowed ksh to used non-compliant implementations of poll(3). This
  feature test fails on POSIX-compliant systems with int-conversion
  errors, which is the expected result. However, the flawed C99
  instrumentation used by Red Hat marks the entire build as a failure
  because of the fact those errors exist at all, even though the
  build proceeds as normal. Fortunately, the operating systems
  supported by ksh (afaik) all have proper, POSIX-compliant versions
  of poll() and thus don't need this test. To work around the flawed
  instrumentation, the unnecessary test has been jettisoned.

Note that the changes in this commit *do not* fix all the errors.
The ones that remain are caused by the incomplete exception list in
redhat-rpm-config, which doesn't include the necessary functions
ksh tests for. That must be fixed in redhat-rpm-config, not ksh.
A patch which adds the required exceptions to the Lua script can
be found in the following comment:
https://github.com/ksh93/ksh/issues/664#issuecomment-1608870665

After applying the patch, the Lua script (which can be run via
/usr/lib/rpm/redhat/report-gcc-errors.sh) will ignore the
implicit function errors that are expected to occur in the feature
tests, which (after fixing conf.tab and _lib_poll) should allow
the build to succeed with Red Hat's instrumentation.