Experimental commit with changes
I will test this commit on my various VMs after pushing. The rename in
signal.c was put there after I saw implicit-function warnings, but that
appears to have actually been caused by the undefs I overlooked.
In any case I don't *think* Haiku needs an extra if directive in map.c
because IIRC signal() is just a macro in the Haiku headers (to be
confirmed or deconfirmed with imminent testing).