Fix typo
On a side note, I haven't had the chance develop local(1) further
because of a lack of time. In any case the current scoping bug is
evading a fix. The current approach needs significant changes,
as it has no backward compatibility when POSIX functions
are nested in ksh functions (the normal ksh93 behavior is to have
POSIX functions inherit the ksh function's full scope and use it
directly, which necessarily includes the ksh function's statically
scoped local variables. The current approach doesn't allow for this
as POSIX functions have their own scope for the local builtin. Backward
compatibility is currently achieved by having typeset modify the global
scope directly, but that's not really the correct thing to do. I have
tried altering typeset to modify the parent ksh function's scope
directly, but such changes cannot be seen in the POSIX function itself
after the variable in the parent function was changed, as the change
only happens in the parent scope and not the current local scope (which
defeats the whole purpose). To attempt to fix that, I then tried linking
the parent scope to the local scope, but so far that just breaks the
local scope (the cdt API lends itself very poorly to dynamic scoping).