Fix non-functioning globcasedetect on Linux (re: 725eef35)
The referenced commit removed the probe that detects the sys/ioctl.h
header under the false assumtion _sys_ioctl would be sufficient. In
reality this change causes iffe to fail to detect sys/ioctl.h on Linux,
which breaks the functionality of the globcasedetect option. Reproducer
using the 'EFI' folder on a UEFI boot partition:
   $ ls /boot/ef*
   ls: cannot access '/boot/ef*': No such file or directory

- Restore the _hdr_sys_ioctl probe to fix compatibility with ext4
  casefold and FAT32.