Fix undefined symbol build error (re: e9278fa)
The aforementioned commit exposed a build error that can occur when
mamake attempts to use ksh as the shell for invoking exec scripts:
  # src/lib/libast/Mamfile: 4796-4801: make ${INSTALLROOT}/dyn/lib/libast.so
  + LDFLAGS='  -Wl,-no-as-needed'
  + export LDFLAGS
  + sed 1d ast.req
  + dylink -m ast -v 6.0 -p lib -s .so state.o opendir.o --redacted object file list--
  [dylink] + cc -Wl,-no-as-needed -o /home/johno/GitRepos/KornShell/ksh/arch/linux.i386-64/dyn/lib/libast.so.6.0 --redacted--
  # src/lib/libast/Mamfile: 4809-4811: make ${INSTALLROOT}/lib/mam
  sh: symbol lookup error: /usr/lib/libcmd.so.2: undefined symbol: _ast_getpgrp
  mamake [lib/libast]: *** exit code 127 making /home/johno/GitRepos/KornShell/ksh/arch/linux.i386-64/lib/mam
  mamake: *** exit code 1 making lib/libast
  mamake: *** exit code 1 making all
  package: make failed at Sun Jul 21 16:16:04 PDT 2024 in /home/johno/GitRepos/KornShell/ksh/arch/linux.i386-64
This error is caused by mamake using the system's installed
ksh binary, which is dynamically linked to its own libast.so.
bin/package exports LD_LIBRARY_PATH and mamake uses it, which
causes ksh to try to use the freshly built libast.so produced
by dylink. This libast.so is binary incompatible with the system's
libcmd.so and by extension /bin/ksh, so the build fails with an
undefined symbol error.

src/cmd/INIT/mamake.c:
- After forking a child process, unset LD_LIBRARY_PATH before invoking
  the shell, as there is no guarantee the shell isn't an older release
  of ksh incompatible with a freshly built libast.so.