simplify nv_getval(); tweak nv_putval() (re: ffbad385, 21b9e509)
This commit should cause no change in behaviour.

In nv_getval(), since vpp is always dereferenced in this function
and never changed, and np->nvalue never changes either, we can
remove one level of pointer dereference by simplifying void **vpp
(pointer to value pointer) to void *vp (value pointer).

This commit also applies a few code legibility tweaks to
nv_putval(): assignment of vpp in an if-then-else; elimination of a
redundant variable (string) and a redundant check for vpp!=NULL
(this is known to be the case there).