mamake: fix %{?} for resume after interrupt (re: 0353bf2f, f5d5318a)
If the user interrupts the build while a target with multiple
prerequisites is being generated, such as libast.a, the contents of
%{?} is incorrect when restarting and resuming the build, because
it only tracks targets updated during the current mamake run.

If an interrupt occurs while building a static library (*.a) and
the build is then resumed, this currently leads to the *.o targets
made before the interrupt not being updated in a pre-existing *.a
library (see use of %{?} in src/cmd/INIT/include/link_ar.mam).

The fix: compare the timestamps of already-built prerequisites to
the timestamp of the parent rule, and add each such prerequisite to
%{?} if it is newer than the parent target.

(This changes the behaviour of %{?} for rules with the 'virtual'
attribute, because those are not bound to files and always start
out with a timestamp of zero. But that's okay because %{?} has
never been used within them.)