src/cmd/ksh93/tests/substring.sh: Add a test to catch bug-833
src/cmd/ksh93/sh/macro.c: Search for phi: bug-833
In string substitution from the end i.e ${string%%pattern} when pattern is
not a pattern but a simple string the original code is apparently broken,
the succession of calls from varsub() to strngrpmatch() make no sense.

This fix is a band-aid, trying not to upset to original code much, it simply
detect that pattern is a pure string (non-pattern) and in that case fall back to
the % replacement instead of %% which are identical.