Fix field separator for $* in here-document
Reproducer:

    IFS=/
    set a b c
    cat <<EOF
    $*
    EOF

Actual output: a b c
Expected output: a/b/c

In any context in which field splitting is not possible, POSIX
specifies that $* uses the first character of $IFS as the output
field separator.[*] Here-documents are such a context.

Analysis: In macro.c, there are three places where IFS is
initialised - in sh_mactrim, sh_macexpand and in sh_machere. The
latter function deals with here-documents. IFS is initialised by
pointing ifsp to the value of IFSNOD and setting ifs to the first
byte of this value. However, in sh_machere, ifs is hardcoded to a
space. This is the bug.

src/cmd/ksh93/sh/macro.c:
- Add setup_ifs function to centralise this logic.
- Call setup_ifs in the three places mentioned, replacing the
  duplicative old code (including the buggy version).

[*] https://pubs.opengroup.org/onlinepubs/9799919799/utilities/V3_chap02.html#tag_19_05_02