libast: rm Inf/NaN binary representation detection (re: a3f4ef7a)
The features/float iffe script contains a large test to detect the
memory representations of the infinity (Inf) and not-a-number (NaN)
values for float, double, and long double.

In the referenced commit, the use of these detected binary patterns
for comparing a value against Inf or NaN was disabled because it is
invalid; the internal representation may contain uninitialised
garbage bits.

The only other use of these detected bit patterns was in sftrtof.h
which is used to generate functions to convert ASCII text to all
types of floating point values. Any garbarge bits are irrelevant
then so this use should be valid.

But it seems unnecessary. The C99 standard provides INFINITY and
NAN constants. Even on systems with pre-C99 system headers that we
still support (like Solaris 9), these values are easy to generate
at compile time: 1.0 / 0.0 should yield infinity and 0.0 / 0.0
should yield not-a-number.

(C99 and POSIX say that not all implementations support NAN, in
which case they do not have NAN defined. I don't know of any such
implementations that ksh runs on. If there are any, hopefully the
0.0 / 0.0 fallback constant will yield some sensible result.)

src/lib/libast/features/float:
- Add fallbacks for INFINITY and NAN. To be fully C99 compliant,
  ensure these are of type float by using the F suffix.
- Remove the large text to detect Inf and NaN bit patterns.

src/lib/libast/sfio/sfstrtof.h:
- Use standard INFINITY and NAN instead of the generated values.

All other changed files:
- Remove variables storing the detected values and related macros.