iffe: revert avoiding rerun of tests (re: d01798f5)
It is actually not correct to decide not to re-run a test merely on
the basis of the timestamp of the test results being up to date.
Some tests depend on other generated headers which may be out of
date, and that might influence the test results.

The last commit (b5747217) was an example. AFter removing the tests
for isnanl and isinfl from the float feature test (generating
ast_float.h), the ksh93 math.sh test results were invalidated on
some systems (at least Android and Haiku) that don't define these
in their system headers. The math.sh test properly depends on
ast_float.h in the Mamfile, but was not rerun because of the
timestamp check added in the referenced commit. So the commit
changed a comment in math.tab to force a rerun. That should not
have been necessary.

This commit removes that timestmap check, fixing the problem. This
also causes test reruns that may not be necessary (because most
tests don't depend on other generated headers), but now that we've
got parallel building, I don't care as much.