Offer all cmds for dynamic libcmd (re: 16c53b0b, 41dfd03a, 35d303f0)
Most people use ksh linked against the static copies of libcmd,
libshell and libast, so the binary is large. To avoid excessive
bloat, only a selection of the built-in commands from libcmd is
included in that binary, though all of them are built.

But when linking against the dynamic libcmd library (on supported
systms), this makes no sense. The commands are in libcmd anyway,
whether ksh links to them or not. You can even link to them at
runtime, e.g., "builtin -f cmd rm" will give you the libcmd
built-in version of rm that is not enabled by default.

There's no reason not to include all the path-bound built-ins by
default when using dynamic libcmd, as it doesn't affect the size of
any of the binaries. Of course they remain bound to /opt/ast/bin so
they only work if that is in the PATH (or after 'builtin
<commandname>'). A simple 'builtin' outputs the full list.

src/cmd/ksh93/data/builtins.c:
- Include all built-ins not only if SHOPT_ALL_LIBCMD is enabled,
  but also if SHOPT_DYNAMIC is enabled. (As of the first referenced
  commit, SHOPT_DYNAMIC is automatically enabled when linking
  against the dynamic cmd/shell/ast libraries).
- To include all built-ins, simply include <cmdlist.h> which was
  generated by libcmd/Mamfile, instead of manually listing them.