libast: cleanup: remove ast.pad bizarritude
TL;DR: WTF, AT&T?

This trailing 'pad' member of the _Ast_info_ struct (ast.pad) has
had me scratching my head for years:

	char		pad[936 - sizeof(void*)];

That size calculation went through many different iterations in the
ast-open-archive history. This is the 93u+ 2012-08-01 version.

If it was meant to ensure that the ast struct has a constant size
for some ABI compatibility purpose or other, then all of those
calculation iterations (including this one) are *wrong*, because
they are all based on a constant number from which various sizeof
results are subtracted at compile time -- but many more of the
added/removed preceding members vary in size from system to system.
Worse, this approach ignores system-defined alignment completely.

But I have no idea what else it could be good for. It looks like a
useless waste of memory.

It gets weirder. The padding space is actually used by one thing:
the Shift-JIS workaround in setlocale.c. It uses that space in a
very ugly and hacky way for two mbstate_t variables, which are
stored at the end of the padding space like this:

  #define mb_state_zero  \
        ((mbstate_t*)&ast.pad[sizeof(ast.pad)-2*sizeof(mbstate_t)])
  #define mb_state \
        ((mbstate_t*)&ast.pad[sizeof(ast.pad)-sizeof(mbstate_t)])

This looks like undefined behaviour to me. Since mbstate_t is an
opaque and implementation-defined type, it's dumb luck the ast.pad
space was big enough for both of them on all systems (or was it?).
Plus, this approach doesn't align them properly.

Not to mention that calling something padding and then using it for
something is just wrong.

I'm calling it. This doesn't have some esoteric purpose. It's BS.

src/lib/libast/include/ast_std.h:
- Remove ast.pad.

src/lib/libast/comp/setlocale.c:
- Use two normal static variables for the Shift-JiS workaround. I
  know, right? How sensible.