libast: utf8_mbtowc(): Fix utf8tab to match new type (re: cb97d3c9)
Given this:

	if ((m = utf8tab[*sp]) > 0)

it was not correct to change the type of m to size_t because that's
an unsigned type, while utf8tab[] contains two final entries of -1.

However, since utf8_mbtowc otherwise treats -1 exactly the same way
as 0, and nothing else uses that table, it's enough just to change
those values to 0.