Fix libast _lib_select test on illumos under clang
This bug exists on the dev branch. When ksh is compiled on illumos with
clang, the libast _lib_select test will fail and subsequently many io.sh
test failures then occur. This is caused by the FD_ZERO macro calling
memset implicitly, which is invalid in C99. To fix this, a string.h
include directive has been added to the feature test.