test/[: Replace check_toomanyops kludge (re: 31953a5d, ffc0f9a9)
My kludge just wasn't sitting well with me. I added it because I
was reluctant to change the test/[ code. But in the last few
commits, I've been audaciously fixing it, so let's revisit this.

Further debugging revealed that the bug was actually caused by
lines 285 and 286 in expr(), here:

285:		if(flag==0)
286:			break;
287:		errormsg(SH_DICT,ERROR_exit(2),e_badsyntax);

That 'break' simply caused excessive arguments to be ignored,
instead of the e_badsyntax error being thrown, unless expr was
processing parentheses or -a/-o. If we remove those lines and the
kludge, the bug stays fixed.

I did a little archaeological research. Those two lines were not
there in ksh86, but were present in ksh88. So the bug was
introduced sometime between 1986 and 1988. It's unknown what those
two lines might have fixed at the time, but it's now known they
cause breakage, so this commit disables them for the POSIX mode.
(On the dev branch, they are removed altogether.)

Removing those lines (and the kludge) on the dev branch did reveal
couple of regression test failures in bracket.sh. The following:
	test -n -a -z
	[ -n -a -z ]
no longer return 0 and are now an error on the dev branch.

But if we look closer at that, we find:
	$ test -n -a -z; echo $?
	0
	$ test -n -a ""; echo $?
	0
which shows that it was broken anyway. Ksh was simply processing
the unary -n operator, testing for the nonemptiness of the string
"-a", and ignoring the superfluous "-z" and "" operands. Which
means the check_toomanyops kludge wasn't doing its job here.

POSIXly, this should be a test for the nonemptiness of the first
and third arguments as strings. So that should be fixed for the
POSIX mode. But real shells and external 'test' commands do vary
here, even in their POSIX modes (if any). So I think -n should
continue to be interpreted as a unary -n in ksh93, with the
excessive argument ignored for backward compatibility on the 1.0
branch.

src/cmd/ksh93/bltins/test.c:
- Remove the check_toomanyops kludge function.
- b_test(): Add POSIX-only special casing for 'test foo -a bar' and
  'test foo -o bar', testing directly for the nonemptiness of the
  lefthand and righthand strings.
- expr(): Disable the offending flag==0 test and break statement
  for the POSIX mode.

Re-fixes: https://github.com/ksh93/ksh/issues/739